Tidy up and cover all methods that take a queue
